requirements.txt
anyio==3.6.2; python_full_version >= "3.6.2" and python_version >= "3.7"
appnope==0.1.3; sys_platform == "darwin" and python_version >= "3.8"
asttokens==2.1.0; python_version >= "3.8"
async-timeout==4.0.2; python_version >= "3.6"
backcall==0.2.0; python_version >= "3.8"
cffi==1.15.1; python_version >= "3.6"
click==8.1.3; python_version >= "3.7"
colorama==0.4.6; python_version >= "3.8" and python_full_version < "3.0.0" and sys_platform == "win32" and platform_system == "Windows" or sys_platform == "win32" and python_version >= "3.8" and python_full_version >= "3.7.0" and platform_system == "Windows"
commonmark==0.9.1; python_full_version >= "3.6.3" and python_full_version < "4.0.0" and python_version >= "3.6"
cryptography==38.0.3; python_version >= "3.6"
decorator==5.1.1; python_version >= "3.8"
deprecated==1.2.13; python_version >= "3.6" and python_full_version < "3.0.0" or python_full_version >= "3.4.0" and python_version >= "3.6"
executing==1.2.0; python_version >= "3.8"
fastapi==0.86.0; python_version >= "3.7"
grpcio-tools==1.50.0; python_version >= "3.7"
grpcio==1.50.0; python_version >= "3.7"
h11==0.14.0; python_version >= "3.7"
httptools==0.5.0; python_version >= "3.7" and python_full_version >= "3.5.0"
idna==3.4; python_full_version >= "3.6.2" and python_version >= "3.7"
ipdb==0.13.9; python_version >= "2.7"
ipython==8.6.0; python_version >= "3.8"
jedi==0.18.1; python_version >= "3.8"
jwt==1.3.1; python_version >= "3.6"
matplotlib-inline==0.1.6; python_version >= "3.8"
packaging==21.3; python_version >= "3.6"
parso==0.8.3; python_version >= "3.8"
pexpect==4.8.0; sys_platform != "win32" and python_version >= "3.8"
pickleshare==0.7.5; python_version >= "3.8"
prompt-toolkit==3.0.32; python_full_version >= "3.6.2" and python_version >= "3.8"
protobuf==4.21.9; python_version >= "3.7"
ptyprocess==0.7.0; sys_platform != "win32" and python_version >= "3.8"
pure-eval==0.2.2; python_version >= "3.8"
pycparser==2.21; python_version >= "3.6" and python_full_version < "3.0.0" or python_full_version >= "3.4.0" and python_version >= "3.6"
pydantic==1.10.2; python_version >= "3.7"
pygments==2.13.0; python_full_version >= "3.6.3" and python_full_version < "4.0.0" and python_version >= "3.8"
pyjwt==2.6.0; python_version >= "3.7"
pyparsing==3.0.9; python_full_version >= "3.6.8" and python_version >= "3.6"
python-dotenv==0.21.0; python_version >= "3.7"
pyyaml==6.0; python_version >= "3.7"
redis==4.3.4; python_version >= "3.6"
rich==12.6.0; python_full_version >= "3.6.3" and python_full_version < "4.0.0" and python_version >= "3.6"
shellingham==1.5.0; python_version >= "3.6"
six==1.16.0; python_version >= "3.8" and python_full_version < "3.0.0" or python_full_version >= "3.3.0" and python_version >= "3.8"
sniffio==1.3.0; python_full_version >= "3.6.2" and python_version >= "3.7"
stack-data==0.6.0; python_version >= "3.8"
starlette==0.20.4; python_version >= "3.7"
toml==0.10.2; python_version > "3.6" and python_full_version < "3.0.0" or python_full_version >= "3.3.0" and python_version > "3.6"
traitlets==5.5.0; python_version >= "3.8"
typer==0.7.0; python_version >= "3.6"
typing-extensions==4.4.0; python_version < "3.9" and python_version >= "3.7" and python_full_version >= "3.6.3" and python_full_version < "4.0.0"
uvicorn==0.19.0; python_version >= "3.7"
uvloop==0.17.0; sys_platform != "win32" and sys_platform != "cygwin" and platform_python_implementation != "PyPy" and python_version >= "3.7"
watchfiles==0.18.1; python_version >= "3.7"
wcwidth==0.2.5; python_full_version >= "3.6.2" and python_version >= "3.8"
websockets==10.4; python_version >= "3.7"
wrapt==1.14.1; python_version >= "3.6" and python_full_version < "3.0.0" or python_full_version >= "3.5.0" and python_version >= "3.6"
